+++ storage-proofs/src/circuit/mod.rs
@@ -1,16 +1,10 @@
 mod constraint;
 
-pub mod por;
-
-pub mod apex_commitment;
-pub mod drgporep;
 pub mod kdf;
 pub mod multi_proof;
 pub mod pedersen;
-pub mod porc;
-pub mod ppor;
+pub mod por;
 pub mod rational_post;
-pub mod sloth;
 pub mod stacked;
 pub mod uint64;
 pub mod variables;
