+++ storage-proofs/src/lib.rs
@@ -28,7 +28,6 @@ pub mod example_helper;
 pub mod circuit;
 pub mod compound_proof;
 pub mod crypto;
-pub mod drgporep;
 pub mod drgraph;
 pub mod error;
 pub mod fr32;
@@ -38,7 +37,6 @@ pub mod merklepor;
 pub mod parameter_cache;
 pub mod partitions;
 pub mod piece_inclusion_proof;
-pub mod porc;
 pub mod porep;
 pub mod proof;
 pub mod rational_post;
@@ -46,5 +44,3 @@ pub mod sector;
 pub mod settings;
 pub mod stacked;
 pub mod util;
-
-pub mod vde;
